requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP and text processing
nltk>=3.8.0

# Transformer models
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Explainability
lime>=0.2.0
shap>=0.43.0

# LLM integration
langchain>=0.1.0
openai>=1.0.0

# Configuration
pyyaml>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Web UI
streamlit>=1.29.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Hugging Face datasets
datasets>=2.14.0

# Progress bars
tqdm>=4.66.0
